requirements.txt
pymongo~=4.11
python-dotenv~=1.0.1
fastapi~=0.115.8
torch~=2.6.0
pillow~=11.1.0
torchvision~=0.21.0
ultralytics~=8.3.74
numpy~=2.2.2
opencv-python~=4.11.0.86
uvicorn~=0.34.0
pydantic~=2.10.6
passlib~=1.7.4
typing_extensions~=4.12.2
PyJWT~=2.10.1